NextFreeSlots: if we got the same slot in multiple requests assume that we have that day complete
This solves the issue with monthly view + user working only once a week. The issue was that we were continously requesting starting with the same partial day, the one day the person was working. And since we never got two different days we never marked a previous day complete.

The solution tough is a workaround because
1. we can not request not full days (eg we can't request slots starting from 12:30)
2. the controller will always return 100 at max, so if there would be more then 100 slots on a single day (eg. because multiple calendars with 10-20 minute length slots) we could never page over

So the fix right now is to assume that its unlikely that we got the same slot twice and that means we will not have a full day worth of slots.